requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.3
gradio>=3.50.2
moviepy>=1.0.3
Pillow>=9.5.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
numpy>=1.24.3

# AI Models
sentencepiece>=0.1.99  # Required for many models
tokenizers>=0.13.3
optimum>=1.8.7  # Model optimization
bitsandbytes>=0.39.0  # 8-bit quantization

# Primary Models
git+https://github.com/huggingface/peft  # Parameter-Efficient Fine-Tuning
einops>=0.6.1  # Required for some vision models

# Fallback Models
protobuf>=3.20.0  # Required for TensorFlow
tensorflow>=2.12.0  # Required for some fallback models
onnxruntime>=1.15.0  # Optimized inference
safetensors>=0.3.1  # Safer model loading

# Asset Management
pytube>=15.0.0  # YouTube video download
yt-dlp>=2023.3.4  # Backup video downloader
pexels-api>=1.0.1
pixabay-python>=2.1
python-unsplash>=1.1.0
freemusicarchive>=0.2.0

# Caching and Storage
diskcache>=5.6.1
requests-cache>=1.0.1
aiofiles>=23.1.0  # Async file operations
fsspec>=2023.5.0  # Filesystem abstraction

# Error Handling and Monitoring
tenacity>=8.2.2  # Retry handling
prometheus-client>=0.17.0  # Metrics
APScheduler>=3.10.1  # Scheduled tasks

# Development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
